requirements.txt
albumentations==1.3.1
coloredlogs==15.0.1
coremltools==7.2
fire==0.5.0
flash_attn==2.5.9.post1
got10k==0.1.3
imageio==2.20.0
kornia==0.7.0
matplotlib==3.7.3
mobile_cv.egg==info
numpy==1.21.6
numpy==1.17.4
omegaconf==2.1.2
onnxruntime==1.15.0
opencv_contrib_python==4.7.0.72
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
packaging==20.3
pandas==1.5.3
Pillow==7.0.0
Pillow==10.4.0
pycocotools==2.0.4
pycolmap==0.6.1
pytorch_lightning==1.5.0
pytorch_toolbelt==0.6.3
PyYAML==6.0.1
scikit_learn==1.0.2
setuptools==67.8.0
setuptools==45.2.0
thop==0.1.1.post2209072238
torchmetrics==0.8.0
torchvision==0.12.0+cu113
tqdm==4.66.1
